refactor apollo client init
